services/audit-service/requirements.txt
aiokafka>=0.10.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
